backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis
qrcode
Pillow
weasyprint
openpyxl
python-dotenv
pytest
httpx
pytest-asyncio
